requirements.txt
blinker==1.9.0
boto3==1.42.1
botocore==1.42.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
dash==3.3.0
dash-bootstrap-components==2.0.4
duckdb==1.4.2
flask==3.1.2
idna==3.11
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
jmespath==1.0.1
markupsafe==3.0.3
narwhals==2.13.0
nest-asyncio==1.6.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
plotly==6.5.0
pyarrow==22.0.0
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
retrying==1.4.2
rpy2-rinterface==3.6.3
rpy2-robjects==3.6.3
s3transfer==0.16.0
schedule==1.2.2
setuptools==80.9.0
six==1.17.0
typing-extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
werkzeug==3.1.4
zipp==3.23.0
